requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
datasets==2.16.1
dill==0.3.7
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.20.3
hydra-core==1.3.2
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
lit==17.0.6
lxml==5.1.0
MarkupSafe==2.1.4
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.37.1
triton==2.0.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
xxhash==3.4.1
yarl==1.9.4
